Created Mood Calendar Componenet
